+++ pkgs/jnigen/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.11.1-wip
 
 - Fixed an issue where inheriting a generic class could generate incorrect code.
+- No longer generating constructors for abstract classes.
+- No longer generating `protected` elements.
 
 ## 0.11.0